src/main/java/com/kensuka/bills/repository/entity/Movement.java
package com.kensuka.bills.repository.entity;

import org.hibernate.validator.constraints.NotEmpty;
import org.springframework.data.annotation.Id;

import javax.validation.constraints.NotNull;
import java.math.BigDecimal;
import java.time.LocalDate;

/**
 * Created by kmukai on 10/7/2016.
 */
public class Movement {
    @Id
    private String id;
    @NotEmpty(message = "Required field.")
    private String description;
    private String type;
    @NotNull(message = "Required field.")
    private BigDecimal amount;
    @NotNull(message = "Required field.")
    private LocalDate date;

    public String getType() {
        return type;
    }

    public void setType(String type) {
        this.type = type;
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public BigDecimal getAmount() {
        return amount;
    }

    public void setAmount(BigDecimal amount) {
        this.amount = amount;
    }

    public LocalDate getDate() {
        return date;
    }

    public void setDate(LocalDate date) {
        this.date = date;
    }
}
